pyproject.toml
[project]
name = "multi-modal-enterprise-knowledge-synthesis-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.8"
dependencies = [
    "aiofiles>=25.1.0",
    "bcrypt>=5.0.0",
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.0",
    "easyocr>=1.7.2",
    "fastapi>=0.129.0",
    "google-genai>=1.63.0",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.12",
    "langchain-huggingface>=1.2.0",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.0.8",
    "lxml>=6.0.2",
    "markdown>=3.10.2",
    "matplotlib>=3.10.8",
    "nltk>=3.9.2",
    "numpy>=2.4.2",
    "olefile>=0.47",
    "openai>=2.21.0",
    "opencv-python-headless>=4.13.0.92",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "pdf2image>=1.17.0",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.11.0",
    "pymongo>=4.16.0",
    "pymupdf>=1.27.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "python-pptx>=1.0.2",
    "python-socketio>=5.16.1",
    "sentence-transformers>=5.2.2",
    "tabulate>=0.9.0",
    "tavily-python>=0.7.21",
    "tiktoken>=0.12.0",
    "torch>=2.10.0",
    "uvicorn[standard]>=0.40.0",
    "wordcloud>=1.9.6",
    "xlrd>=2.0.2",
]
